reimplement: SHC_3BB0A8C1_0x0045FDC0 match 16-bit border arithmetic
The original computes the "unknown2" line-layout field entirely in
16 bit:

    mov ax, word ptr [esi + eax*2]      ; rightBorder
    sub ax, word ptr [esp + 0x14]       ; leftPosition (low word)
    mov word ptr [esi + ebp*2 + 0x62c0], ax

Because leftPosition is an int, the reimplementation promoted the
expression to 32 bit and emitted a sign-extending load plus a 32-bit
subtraction at all five assignment sites. Narrowing leftPosition to
short restores the original's 16-bit load/subtract/store sequence.

The result is unchanged: the difference is only in which width the
truncation to the short field happens in.

+++ src/OpenSHC/Text/TextEditorState/processHelpRichTextTokens.cpp
@@ -219,7 +219,7 @@ namespace Text {
                             local_f0 = FALSE;
                             if (local_120) {
                                 this->lineLayoutTable[lineY].unknown2
-                                    = this->lineLayoutTable[lineY].rightBorder - leftPosition;
+                                    = this->lineLayoutTable[lineY].rightBorder - (short)leftPosition;
                                 local_120 = centreActive;
                             }
                             do {
@@ -438,7 +438,8 @@ namespace Text {
                 }
             case Enums::HTT_NEWPARAGRAPH:
                 if (local_120) {
-                    this->lineLayoutTable[lineY].unknown2 = this->lineLayoutTable[lineY].rightBorder - leftPosition;
+                    this->lineLayoutTable[lineY].unknown2
+                        = this->lineLayoutTable[lineY].rightBorder - (short)leftPosition;
                     local_120 = centreActive;
                 }
                 if (this->useAlternateHelpTab && iVar5 == this->activeHelpHotspotIndex) {
@@ -516,7 +517,7 @@ namespace Text {
                         local_f0 = FALSE;
                         if (local_120) {
                             this->lineLayoutTable[lineY].unknown2
-                                = this->lineLayoutTable[lineY].rightBorder - leftPosition;
+                                = this->lineLayoutTable[lineY].rightBorder - (short)leftPosition;
                             local_120 = centreActive;
                         }
                         do {
@@ -603,7 +604,8 @@ namespace Text {
                 if (this->lineLayoutTable[lineY].rightBorder < local_f4) {
                     local_f0 = FALSE;
                     if (local_120) {
-                        this->lineLayoutTable[lineY].unknown2 = this->lineLayoutTable[lineY].rightBorder - leftPosition;
+                        this->lineLayoutTable[lineY].unknown2
+                            = this->lineLayoutTable[lineY].rightBorder - (short)leftPosition;
                         local_120 = centreActive;
                     }
                     lineY += 1 + currentFontLineHeight;
@@ -638,7 +640,7 @@ namespace Text {
                                 iVar5 = local_11c;
                                 if (local_120) {
                                     this->lineLayoutTable[lineY].unknown2
-                                        = this->lineLayoutTable[lineY].rightBorder - leftPosition;
+                                        = this->lineLayoutTable[lineY].rightBorder - (short)leftPosition;
                                     local_120 = centreActive;
                                 }
                                 lineY += currentFontLineHeight + 1;
